DTP-78 Moved interface for adapter to the presentation folder
