+++ src/VPA.Client.VisualStudio.Extension.VSIX.Tests/DetectorResultCollectionToTreeViewAdapterTests.cs
@@ -1,5 +1,5 @@
 using System.Windows.Controls;
-using VPA.Client.VisualStudio.Extension.VSIX.Adapters.Presentation;
+using VPA.Client.VisualStudio.Extension.VSIX.Adapters;
 using VPA.Domain.Models;
 
 namespace VPA.Client.VisualStudio.Extension.VSIX.Tests
