+++ src/VPA.Client.VisualStudio.Extension.VSIX/Adapters/IDetectorResultCollectionToTreeViewAdapter.cs
@@ -1,7 +1,7 @@
 ﻿using System.Windows.Controls;
 using VPA.Domain.Models;
 
-namespace VPA.Common.Adapters.Interfaces.Presentation
+namespace VPA.Client.VisualStudio.Extension.VSIX.Adapters
 {
 	public interface IDetectorResultCollectionToTreeViewAdapter
 	{
