+++ src/VPA.Client.VisualStudio.Extension.VSIX/ToolWindows/DesignPatternWindowControl.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="VPA.Client.VisualStudio.Extension.VSIX.DesignPatternWindowControl"
+﻿<UserControl x:Class="VPA.Client.VisualStudio.Extension.VSIX.ToolWindows.DesignPatternWindowControl"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
