+++ src/VPA.Client.VisualStudio.Extension.Vsix/VPA.Client.VisualStudio.Extension.Vsix.csproj
@@ -45,7 +45,8 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="Adapters\Presentation\DetectorResultCollectionToTreeViewToTreeViewAdapter.cs" />
+    <Compile Include="Adapters\DetectorResultCollectionToTreeViewToTreeViewAdapter.cs" />
+    <Compile Include="Adapters\IDetectorResultCollectionToTreeViewAdapter.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Commands\OpenDesignPatternWindow.cs" />
     <Compile Include="VPA.Client.VisualStudio.Extension.VSIXPackage.cs" />
@@ -130,6 +131,7 @@
       <Name>VPA.Usecases</Name>
     </ProjectReference>
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="$(VSToolsPath)\VSSDK\Microsoft.VsSDK.targets" Condition="'$(VSToolsPath)' != ''" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
